refactor(weekgrid): the grid's machinery becomes a hook
The running board around the geometry — the window the reader widened,
the zoom and per-column widths they chose and where those are kept, the
measurement the layers are placed by, the open-on-today scroll, and the
two hit tests a pointer gesture asks — was two hundred lines inside
ProjectBoard, and a second board holding a week grid would have had to
write all of it again.

useWeekGrid holds it, and knows nothing about what a column is: it
counts them, measures them and answers with an index. ProjectBoard is
three hundred lines shorter and draws exactly what it drew.
